requirements.txt
pika
requests
python-dotenv
schedule
